Added: Exception Handler and extended Model

+++ src/main/java/com/bridgelabz/repository/IEmployeeRepository.java
@@ -3,6 +3,6 @@ package com.bridgelabz.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.bridgelabz.model.Employee;
 
-public interface IEmployeeRepository extends JpaRepository<Employee, Long> {
+public interface IEmployeeRepository extends JpaRepository<Employee, Integer> {
 
 }
